# docs/api-documentation.yaml
openapi: 3.0.3
info:
  title: Voke E-commerce API
  description: |
    Complete API for the Voke electronic store.
    
    This API provides endpoints to manage products, categories, shopping cart, and user authentication.
    
    ## Main Features:
    - Complete electronic product management
    - Advanced filtering by categories, prices, and brands
    - Product search functionality
    - Result pagination
    - Featured products and offers
    - Shopping cart with session support
    - JWT user authentication
    - Interactive Swagger documentation
    
    ## Authentication
    The API uses JWT (JSON Web Tokens) for authentication. For protected endpoints,
    include the token in the Authorization header:
    ```
    Authorization: Bearer <your_jwt_token>
    ```
    
    ## Shopping Cart
    The cart supports both anonymous users (using session_id) and authenticated
    users (using user_id). When logging in, the cart is automatically migrated.
    
    ## Base URL
    - **Development:** `http://localhost:3001/api`
    - **Production:** `https://tu-app-voke-backend-7da6ed58e5fc.herokuapp.com/api`
  version: 1.0.0
  contact:
    name: Voke Team
    email: support@voke.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://tu-app-voke-backend-7da6ed58e5fc.herokuapp.com/api
    description: Production Server (Heroku)
  - url: http://localhost:3001/api
    description: Development Server
paths:
...
  /cart:
    get:
      summary: Get shopping cart
      description: Retorna todos los productos en el carrito del usuario o sesión
      tags:
        - Cart
      parameters:
        - name: x-user-id
          in: header
          description: User ID (optional)
          required: false
          schema:
            type: string
            example: user123
        - name: x-session-id
          in: header
          description: Session ID (optional)
          required: false
          schema:
            type: string
            example: session456
        - name: user_id
          in: query
          description: User ID (alternativo)
          required: false
          schema:
            type: string
        - name: session_id
          in: query
          description: Session ID (alternativo)
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Cart retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/CartItem'
                  count:
                    type: integer
                    example: 3
        '400':
          description: Invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /cart/summary:
    get:
      summary: Get cart summary
      description: Returns cart summary with totals
      tags:
        - Cart
      parameters:
        - name: x-user-id
          in: header
          description: User ID
          required: false
          schema:
            type: string
        - name: x-session-id
          in: header
          description: Session ID
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Cart summary retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/CartSummary'
        '400':
          description: Invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /cart/add:
    post:
      summary: Add product to cart
      description: Adds a product to the cart o actualiza la cantidad si ya existe
      tags:
        - Cart
      parameters:
        - name: x-user-id
          in: header
          description: User ID
          required: false
          schema:
            type: string
        - name: x-session-id
          in: header
          description: Session ID
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                product_id:
                  type: string
                  format: uuid
                  description: Product ID
                  example: 123e4567-e89b-12d3-a456-426614174000
                quantity:
                  type: integer
                  minimum: 1
                  description: Cantidad a agregar
                  example: 2
              required:
                - product_id
      responses:
        '201':
          description: Product added to cart successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Product added to cart
                  data:
                    $ref: '#/components/schemas/CartItem'
        '400':
          description: Invalid data o stock insuficiente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /cart/update/{id}:
    put:
      summary: Update cart quantity
      description: Updates product quantity en el carrito
      tags:
        - Cart
      parameters:
        - name: id
          in: path
          required: true
          description: ID del item del carrito
          schema:
            type: string
            format: uuid
        - name: x-user-id
          in: header
          description: User ID
          required: false
          schema:
            type: string
        - name: x-session-id
          in: header
          description: Session ID
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                quantity:
                  type: integer
                  minimum: 0
                  description: Nueva cantidad (0 para eliminar)
                  example: 3
              required:
                - quantity
      responses:
        '200':
          description: Cantidad actualizada successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Cantidad actualizada
                  data:
                    $ref: '#/components/schemas/CartItem'
        '400':
          description: Invalid data o stock insuficiente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /cart/remove/{id}:
    delete:
      summary: Remove product from cart
      description: Deletes a product específico del carrito
      tags:
        - Cart
      parameters:
        - name: id
          in: path
          required: true
          description: ID del item del carrito
          schema:
            type: string
            format: uuid
        - name: x-user-id
          in: header
          description: User ID
          required: false
          schema:
            type: string
        - name: x-session-id
          in: header
          description: Session ID
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Product deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Product removed from cart
        '400':
          description: Invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /cart/clear:
    delete:
      summary: Clear cart
      description: Elimina todos los productos del carrito
      tags:
        - Cart
      parameters:
        - name: x-user-id
          in: header
          description: User ID
          required: false
          schema:
            type: string
        - name: x-session-id
          in: header
          description: Session ID
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Cart cleared successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Cart cleared successfully
        '400':
          description: Invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /cart/migrate:
    post:
      summary: Migrate session cart to user
      description: Migrates cart products de una sesión anónima a un usuario autenticado
      tags:
        - Cart
      parameters:
        - name: Authorization
          in: header
          description: JWT token del usuario autenticado
          required: true
          schema:
            type: string
            example: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                session_id:
                  type: string
                  description: ID de la sesión anónima
                  example: session456
              required:
                - session_id
      responses:
        '200':
          description: Cart migrated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Cart migrated successfully
                  data:
                    type: object
                    properties:
                      migratedItems:
                        type: integer
                        example: 3
                      totalValue:
                        type: number
                        format: float
                        example: 12899.97
        '400':
          description: Invalid data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # =====================================================
  # ENDPOINTS DE AUTENTICACIÓN
  # =====================================================
...
components:
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique product ID
          example: 123e4567-e89b-12d3-a456-426614174000
        title:
          type: string
          description: Título del producto
          example: iPhone 13 128GB - Azul
        brand:
          type: string
          description: Product brand
          example: Apple
        category:
          type: string
          description: Product category
          enum: [Smartphone, Tablet, Notebook, Electronics]
          example: Smartphone
        original_price:
          type: number
          format: float
          description: Price original del producto
          example: 4999.99
        current_price:
          type: number
          format: float
          description: Price actual del producto
          example: 4299.99
        discount:
          type: integer
          description: Porcentaje de descuento
          minimum: 0
          maximum: 100
          example: 14
        installments:
          type: object
          description: Información de cuotas
          properties:
            times:
              type: string
              example: 12x
            value:
              type: string
              example: 358,33
        image:
          type: string
          description: Image URL del producto
          example: https://example.com/image.jpg
        description:
          type: string
          description: Product description
          example: iPhone 13 com chip A15 Bionic, sistema de câmera dupla avançado
        stock:
          type: integer
          description: Cantidad en stock
          minimum: 0
          example: 20
        is_active:
          type: boolean
          description: Si el producto está activo
          example: true
        is_featured:
          type: boolean
          description: If the product is featured
          example: true
        is_offer:
          type: boolean
          description: If the product is on sale
          example: true
        is_best_seller:
          type: boolean
          description: Si el producto es más vendido
          example: true
        condition:
          type: string
          description: Product condition
          enum: [Novo, Seminovo, Used]
          example: Novo
        warranty:
          type: string
          description: Garantía del producto
          example: 1 ano
        specifications:
          type: object
          description: Especificaciones técnicas del producto
          example:
            processador: A15 Bionic
            armazenamento: 128GB
            camera: 12MP dupla
            tela: 6.1 polegadas
        created_at:
          type: string
          format: date-time
          description: Creation date
          example: 2025-08-22T16:33:53.03547+00:00
        updated_at:
          type: string
          format: date-time
          description: Last update date
          example: 2025-08-22T16:33:53.03547+00:00
      required:
        - id
        - title
        - category
        - current_price
        - is_active
...
    CartItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: ID único del item del carrito
          example: cart123e4567-e89b-12d3-a456-426614174000
        user_id:
          type: string
          description: User ID (si está logueado)
          example: user123
        session_id:
          type: string
          description: Session ID (si es anónimo)
          example: session456
        product_id:
          type: string
          format: uuid
          description: Product ID
          example: 123e4567-e89b-12d3-a456-426614174000
        quantity:
          type: integer
          minimum: 1
          description: Product quantity
          example: 2
        price:
          type: number
          format: float
          description: Product price al momento de agregarlo
          example: 4299.99
        created_at:
          type: string
          format: date-time
          description: Creation date
          example: 2025-08-22T16:33:53.03547+00:00
        updated_at:
          type: string
          format: date-time
          description: Last update date
          example: 2025-08-22T16:33:53.03547+00:00
        products:
          $ref: '#/components/schemas/Product'
      required:
        - id
        - product_id
        - quantity
        - price
    CartSummary:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/CartItem'
          description: Lista de productos en el carrito
        itemCount:
          type: integer
          description: Number of unique products
          example: 3
        totalQuantity:
          type: integer
          description: Cantidad total de productos
          example: 5
        subtotal:
          type: number
          format: float
          description: Subtotal sin descuentos
          example: 12999.95
        totalDiscount:
          type: number
          format: float
          description: Total de descuentos aplicados
          example: 1500.00
        total:
          type: number
          format: float
          description: Total final a pagar
          example: 11499.95
      required:
        - items
        - itemCount
        - totalQuantity
        - subtotal
        - totalDiscount
        - total
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          description: Mensaje de error
          example: Internal server error
        error:
          type: string
          description: Detalles del error
          example: Database connection failed
      required:
        - success
        - message
...